+++ src/WireMock.Net/Util/StringUtils.cs
@@ -8,6 +8,48 @@ namespace WireMock.Util;
 
 internal static class StringUtils
 {
+    private static readonly string[] ValidUriSchemes =
+    {
+        "ftp://",
+        "http://",
+        "https://"
+    };
+
+    private static readonly Func<string, (bool IsConverted, object ConvertedValue)>[] ConversionsFunctions =
+    {
+        s => bool.TryParse(s, out var result) ? (true, result) : (false, s),
+        s => int.TryParse(s, out var result) ? (true, result) : (false, s),
+        s => long.TryParse(s, out var result) ? (true, result) : (false, s),
+        s => double.TryParse(s, out var result) ? (true, result) : (false, s),
+        s => Guid.TryParse(s, out var result) ? (true, result) : (false, s),
+        s => TimeSpan.TryParse(s, out var result) ? (true, result) : (false, s),
+        s => DateTime.TryParse(s, out var result) ? (true, result) : (false, s),
+        s =>
+        {
+            if (ValidUriSchemes.Any(u => s.StartsWith(u, StringComparison.OrdinalIgnoreCase)) &&
+                Uri.TryCreate(s, UriKind.RelativeOrAbsolute, out var uri))
+            {
+                return (true, uri);
+            }
+
+            return (false, s);
+        }
+    };
+
+    public static (bool IsConverted, object ConvertedValue) TryConvertToKnownType(string value)
+    {
+        foreach (var func in ConversionsFunctions)
+        {
+            var result = func(value);
+            if (result.IsConverted)
+            {
+                return result;
+            }
+        }
+
+        return (false, value);
+    }
+
     public static MatchOperator ParseMatchOperator(string? value)
     {
         return value != null && Enum.TryParse<MatchOperator>(value, out var matchOperator)